versione 0

+++ src/main/java/com/generation/the_elder_bean/models/dto/CombattimentoRequest.java
@@ -5,26 +5,26 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 public class CombattimentoRequest
 {
 
-	private long id1;
+	private int id1;
 
-	private long id2;
+	private int id2;
 
-	public long getId1()
+	public int getId1()
 	{
 		return id1;
 	}
 
-	public void setId1(long id1)
+	public void setId1(int id1)
 	{
 		this.id1 = id1;
 	}
 
-	public long getId2()
+	public int getId2()
 	{
 		return id2;
 	}
 
-	public void setId2(long id2)
+	public void setId2(int id2)
 	{
 		this.id2 = id2;
 	}
